+++ CHANGELOG.md
@@ -17,6 +17,7 @@ high state of flux, you're at risk of it changing without notice.
 
 * **Bug Fix**
   * add phantom fields to curried type classes, fix #316 (@gcanti)
+  * fix `Unfoldable.replicateA` signatures (@gcanti)
 
 # 1.0.0